Changes: Added Scoreboard command
Type or commit:
feature

Additional Comments:
-

+++ migrations/20200518190807_discord_servers.js
@@ -2,7 +2,7 @@ exports.up = (knex) => {
 	return knex.schema
 		.createTable('discord_servers', (table) => {
 			table.increments('id');
-			table.integer('server_id').notNullable();
+			table.bigInteger('server_id').notNullable();
 			table.string('server_name').notNullable();
 			table.integer('server_members_no').notNullable();
 		});
